+++ src/routes/Home/style.module.css
@@ -1,0 +1,6 @@
+.flex {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
